requirements.txt
pymongo
requests
python-dotenv
docker
